+++ tests/testthat/resources/targets.R
@@ -1,0 +1,1 @@
+tar_option_set(packages = c("A", "B"))